Use a single instance of a client for Workday
Because the WSDL file is so large (3.2MB), it takes an awful long time
for a newly instantiated LolSoap::Client to parse it (upwards of 2-3
seconds). Instead of creating a new client each time, we create a single
instance and reuse that to speed up execution of both the running app
and the test suite.