Re-instate validations for RM1031
The validations for this framework were slackened before the framework
went live in October because it was discovered that suppliers were
already submitting data with invalid data.

In collaboration with Malcolm Eastoe from the MI Team, we are now
re-instated some essential validations that will not cause major
failures for suppliers.

A small number of existing submissions do fail with these changes,
all for Berendsen UK Limited, which isn't enough to cause concern for
suppliers making valid submissions in the future:

    eae7f01d-581a-4e37-bdfe-9d8f3cbcb956
    639b4ffa-d42b-4d47-a358-f86d56a135f3
    e8a6b9f2-169f-4db2-a336-11e58e112646