Re-instate validations for RM1070
The validations for this framework were slackened before the framework
went live in October because it was discovered that suppliers were
already submitting data with invalid data.

In collaboration with Malcolm Eastoe from the MI Team, we are now
re-instated some essential validations that will not cause major
failures for suppliers.

A small number of existing submissions do fail with these changes, but
not enough to cause concern for suppliers making valid submissions in
the future:

   0e14ad0a-ab58-46b1-af91-f3ebd31a728b
   c6b19476-281c-4ebe-a75f-faac1a047dee
   613af265-c5f0-4735-be2b-647704a853fb
   cd36ba6c-7578-46a6-aaed-40700bfc8c83
   6ea7c012-8825-4faa-a0f5-e043bb6130f5
   090b2125-bac1-45b7-b72f-271d2fb67863
   ce452897-79e8-49e4-b439-c8e3d3ad4519

Malcolm is to go back to the CAM for this framework to discus three
columns that should be validated, but currently would result in a large
number of existing submissions failing validation: "Vehicle CAP Code",
"Customer Support Terms", "CO2 Emissions".