Re-instate validations for 3710
The validations for this framework were slackened before the framework
went live in October because it was discovered that suppliers were
already submitting data with invalid data.

In collaboration with Malcolm Eastoe from the MI Team, we are now
re-instated some essential validations that will not cause major
failures for suppliers.

A small number of existing submissions do fail with these changes, but
not enough to cause concern for suppliers making valid submissions in
the future. One supplier (Volkswagen Financial Services (UK) Ltd) has
consistently missed the "Customer Invoice Number" from their returns.
Malcolm to go back to the CAM and let them know about the need to fill
this in in future.

Malcolm is to go back to the CAM for this framework to discus two
columns that should be validated, but currently would result in a large
number of existing submissions failing validation: "CAP Code", "CO2
Emission Levels".