Re-instate validations for CM/OSG/05/3565
The validations for this framework were slackened before the framework
went live in October because it was discovered that suppliers were
already submitting data with invalid data.

In collaboration with Malcolm Eastoe from the MI Team, we are now
re-instated some essential validations that will not cause major
failures for suppliers.

One existing submission fails with these changes, which isn't enough
to cause concern for suppliers making valid submissions in the future:

   bf9ce182-ded0-48b7-940e-a9ed247dd787