Re-instate validations for RM3754
The validations for this framework were slackened before the framework
went live in October because it was discovered that suppliers were
already submitting data with invalid data.

In collaboration with Malcolm Eastoe from the MI Team, we are now
re-instated some essential validations that will not cause major
failures for suppliers.

A small number of existing submissions do fail with these changes, but
not enough to cause concern for suppliers making valid submissions in
the future.