Re-instate validations for RM3767
The validations for this framework were slackened before the framework
went live in October because it was discovered that suppliers were
already submitting data with invalid data.

In collaboration with Malcolm Eastoe from the MI Team, we are now
re-instated some essential validations that will not cause major
failures for suppliers.

A small number of existing submissions do fail with these changes, but
not enough to cause concern for suppliers making valid submissions in
the future:

  49493aa4-ff1c-4697-a2d6-7c6b38fa8304
  f6a5e94d-20e3-41c3-b81a-c9f750063332
  63b16d3b-cdba-4d81-96a0-8c4108c775aa
  45afb85e-0537-4c54-acbf-8882f0b37654
  5a3e8239-5c69-4455-91c5-4b5572ab2e90
  d80236ef-816d-49b1-b337-2e7365afb662
  db31f1b9-b12c-4c19-b116-0ec298e01bee