Management Charge calculation accepts entry rather that just total_value
We'll need access to the rest of the entry in order to handle
variable management charges, so this changes the value passed in to this
method.

Also, rename 'management_charge' method to 'calculate_management_charge'
which makes the method's purpose more explicit, particularly now it's not
a fixed value per framework.