Add new class to calculate column-based management charges
RM3710 and RM858 apply different management charge rates based on the
contents of the "Spend Code" column in the submitted spreadsheet.

The new `ManagementChargeCalculation::ColumnBased` accepts two
parameters, `column` and `value_to_percentage`:

 - `column` is the column in the spreadsheet in which to look for a
value
 - `value_to_percentage` is a Ruby hash which maps the value (in the
`column` defined above to the amount that should be charged.

Because we generally do case-insensitive validation on the fields used
here, we downcase both the entry column value and the
'value_to_percentage' hash keys.

To maintain compatibility with existing framework definitions, if
`management_charge_rate` is a BigDecimal, treat as a
`ManagementChargeCalculation::FlatRate`.