Show admin users supplier tasks and latest submission status
So that admin users can support supplier users with issues
they need to see the status of the latest submission a supplier
has made.

We've added a date and a time format to display the submission
date and time and the submission month correctly.

The supplier controller and view now have a partial to render the
task and submission based on it's status.

It looks like the submission factories may not be moving submissions
to the correct status so for now we are setting the status on submission
on creation in the test.