(fix) Workday test submission is associated with correct task
Associate submissions with the task created in the test, rather than one
created automatically by FactoryBot (which uses the current date to set
`period_month` and `period_year`).

On February 1st, these tests started to fail because the task under test
was now in a different period to the task associated with the submission.