Submit Revenue Category and Tax Code IDs to Workday
Workday needs us to submit the correct Revenue Category and Tax Code IDs for the relevant framework. This data is retrieved from Workday.

We create a new helper class called "Workday::CommercialAgreements" to retrieve and extract those IDs from Workday, and use them to add it to the Invoice and InvoiceAdjustment XML that is submitted.