(fix) Dependent field inclusion validator bug 2/2
Previously, if the entry data did not contain a value
for the tested field, the validator would crash,
and the submission would get stuck.

This commit fixes that: if the dependent field value
is missing, the validator will not crash.

All the fields validated in a dependent way are currently also
validated for presence, so the missing value will be detected
by the presence validator.