Link to the submission file if there is one
Admin users see a least of tasks and submissions on the supplier view.
This commit adds a link the the submission file, when there is one,
so that an admin user can download the file and troubleshoot any
supplier issues.

Updated the submission factory to include the attached file trait to
test against.

Used rails_blob_url which gets the appropriate file url for the
environment (S3 for dev and local for testing).

Shows the link on the task list on the supplier page.