Make Analytical amounts negative again
Submissions with negative management charge are submitted to Workday as InvoiceAdjustments. In a previous commit, we made a change that InvoiceAdjustments should be set to "Credit: true" and the negative amounts made positive.

It has been requested that while management charges (Extended Amount in Workday) remain positive, total spend (Analytical amount in Workday) is reverted back to negative. This also updates the Line Item Description as desired.