Allow free login as admin in development
To avoid devs needing to set up their own Google credentials, create
a custom OmniAuth DeveloperAdmin strategy. This takes a single email
address, usually set up in .env.development, for example
`ADMIN_EMAILS=my.email@example.com`. When it `applies?`, which is to
say in development and when either `GOOGLE_CLIENT_*` value is missing,
it requires:

- setup from config/initializers/omniauth.rb
- an extra POST route
- a link to the proper strategy on the sign-in button
- skipping CSRF protection for `admin/sessions#create`

This is used over the built-in Developer strategy for two reasons:

1. The built-in developer strategy doesn't give any leeway for
   communicating to the user that the "Email" field must correspond
   to the ADMIN_EMAILS env var, so that requires docs and action
2. The built-in developer strategy has a problem with using an
   `option :name`, which we use in our info_hash and which overrides the
   default route to the strategy, resulting in 404s and
   head-scratching