Say when we need to set ADMIN_EMAIL
The concept of a default email isn't useful – we're only an admin
when there's an email address in an ADMIN_EMAILS list locally.

If that value isn't there, tell developers to set it up. Also, label
the fields in the developer admin form – `#html` won't do that for us.

Lastly, acknowledge that even though most of the time there'll be a
single email address in ADMIN_EMAILS for development, it's meant to be a
list, so allow for that possibility and only use the first one.