(chore) Refactor UserController#create to have clearer exception handling
Initially this controller was refactored when https://rollbar.com/dxw/ccs-reportmi-api/items/65/occurrences/66083571076/ indicated a user had somehow been created without an Auth0 id. However on closer inspection it transpires that the exception was raised on a double-submit of the User#destroy action.

However it's been decided to keep this refactor as the scoping of the exception handling is clearer than in the earlier implementation.