Replace a submission with a no business return
The no business API endpoint accepts a parameter `replacement` in order to replace an existing business return with a no business return.

Note: this is not exposed through the interface yet.

Also note: it is not permitted to replace a no business with another no business, on the grounds that it woudn’t make sense.

PS: Turning off the rubocop rule as agreed with Leeky because:
- it makes code much less readable
- the performance benefits are doubtful: https://github.com/rubocop-hq/rubocop/issues/4052#issuecomment-465865434

Co-authored-by: CristinaRO <cristina@dxw.com>