Document the data warehouse export
This was requested so that admin users can understand what data is
available to them in the data warehouse so that they know what they find
there when the RMI admin interface doesn't show them what they need.
This will go into the service manual, but I figured it would also be
useful to document it here in the application source so it's easily
accessible to developers.