Optimise the GET /tasks/ endpoint
Because of the way that nested resources work in JSONAPI, including
`invoice_total_value` and `order_total_value` on a submission was
causing this to also get calculated when returning the list of tasks,
even though this information is never displayed on the frontend.

We've used JSONAPI's `fields` feature to allow the frontend to request
only a subset of the attributes on the Task model (and its
associations). Unfortunately, the `fields` parameter that gets passed in
from the frontend wasn't getting understood by the JSONAPI library, so
some modification was needed in order for it to get parsed correctly.

Additionally, the ActiveRecord query includes the framework, supplier
and latest_submission models, to reduce the large number of SQL queries
that were previously being run.