Admin can deactivate a supplier from frameworks
Suppliers will occasionally need to be deactivated from a framework so
that they are no longer required to submit management information. This
gives admin users the ability to deactivate a given supplier from their
frameworks. Future commits will a) add the ability to re-activate them,
and b) update task generation so that inactive agreements do not
generate tasks.