Tasks endpoint can optionally return only selected fields
Implement sparse fieldsets so that the frontend can request a subset of
attributes. This means we don't use the processing overhead to generate
calculated attributes that aren't ever used.

The JSON API specification defines the fields parameter as an array
which maps model name to a comma-separated list of attributes, e.g.

```
fields[submissions]=status,submitted_at
```

This required some additional work to deserialise this into a format
that the 'jsonapi-rails' gem could understand.

Co-authored-by: Tekin Suleyman <tekin@tekin.co.uk>