Tasks endpoint avoids N+1 queries
Any associations mentioned in the `include` parameter, are also included
in the ActiveRecord call, to avoid N+1 queries.

Additionally, we always include the `Supplier` model, because the
`Task` model delegates `supplier_name`, which causes additional
queries to be made.

If the API receives a request for either an invalid association or a
nested one, it'll return an HTTP 500 Internal Server Error, and the
`ActiveRecord::AssociationNotFoundError` will appear in the logs.