Remove Rubocop warning
Even though in our case it makes no difference, we either need to add
a &. to the downcase in dependent_field_inclusion_validator or we need
to explicitly exclude this check. There won't be any circumstances in
which downcase could return nil, but let's quieten the warning noise
in the build output.