LateGenerator becomes OverdueUserNotificationList
This is in anticipation of a related/similar (but not quite the same)
Task::AnticipatedUserNotificationList class that will be used to output
CSV for emails to go to users that will have tasks in the next
submission window. Task::LateGenerator didn't feel like it accurately
reflected what the class did, and the problem was compounded when trying
to come up with a name for the newly required class that was similarly
named.