Add rake generate:anticipated_user_task
Generates CSV for all the active users that should get tasks in the
specified period. The underlying code is written such that it can be run
against a future reporting period (i.e. the tasks don't need to exist)
so that the list can be generated ahead of time and scheduled for
sending in GOV.UK Notify the morning of the submission window.

Outputs to STDOUT by default (pipe to a file if you know where it's
going; no standard location is implied)