Respond with latest submission for nochange
Previously, when a user tried to resubmit no business against a task which already had a completed no business return, we did nothing and returned the *first* submission against that task.

This may not have been the latest submission. This commit ensures that we always return the latest submission, rather than a random one.

Co-authored-by: CristinaRO <cristina@dxw.com>