Remove flat-rate charge calculator magic
Removing the special case/default for the flat rate calculator will make
it easier to enhance the calculator to accept a second parameter that
specifies the column in the entry to calculate the charge against. This
is necessary because not all frameworks calculate their charge against
the total value columm.