FlatRate calculator can override calculation field
This makes it possible to override the column that the management charge
calculation is performed against, which is necessary for RM6060 and
other future frameworks.

The the value coming from the `data` hash will be either a string or a
number. To make sure the calculation is performed the same as with
`total_value`, we cast it to a `BigDecimal` using `to_d`, otherwise we
end up with rounding/truncation inconsistency.