Prevent creating new submissions if completed exists
To allow the creation of correction submissions, we want to be able to create submissions against tasks that already have a completed submission.

However, we only want to do this if the correction parameter has been set.

The situation we're trying to prevent is most likely to be double- or multiple-clicking when submitting the new submission form.

Co-authored-by: james <james@abscond.org>