Mark old submissions as replaced
The old submissions are marked as `replaced` in the same transaction as marking the latest submission as `completed`.

This is to ensure that there is a completed submission at all times, even while the latest submission (the correction) is in the process of being validated and reviewed.

Note: it's not possible to arrive in a situation with multiple `completed` submissions through application code, but it is possible through console interventions such as support requests. We will be adding a data migration in the next commit to fix the existing submissions.

Co-authored-by: james <james@abscond.org>