Additional fields with user-defined types
We need to be able to parse AdditionalN fields as fields with a type
other than a primitive (currently only String but will be Decimal, YesNo
etc). So we need to state explicitly that unknown fields always take
a primitive type in order to be able to tell them apart from
Additional fields that are using a PascalCased type to reference a
lookup.