Introduce Task#incomplete? method
This will indicate whether a task is incomplete or not. The spec looks
more complicated than it might need to be, but it's written such that if
a new state is introduced in the future, it will fail without someone
explicitly making the decision as to whether a task in this new state
should be considered incomplete or not. This makes it less likely that
hypothetical "archived" tasks will incorrectly appear in the late
notifications code.