Make hashdiff gem available in production mode
This gem is only meant for use in development mode when testing and
developing FDL using FDL::Validations::Test. But because there is an
`fdl.rake` file that requires that class' file, hashdiff also gets
required. This results in any `bundle exec rake` commands run in
production mode (where bundle install is run with ` --without
development test`) failing with a "cannot load such file -- hashdiff"
error.

For now we make the gem available in production mode, but don't
auto-require it.