Improve have_field matcher failures
The default custom matcher was helpful in that it told you what was
expected – it just didn't tell you what it actually found, so we spent a
bit of time debugging before getting things right. Print a summary of
actual validators found for an attribute, one per line, so that we don't
have to do that any more.