Simply Task#completed_or_latest_scope
The previous query, whilst it worked, relied on binding to the task_id
using the $1 substitution parameter. Although this worked fine in most
scenarios, it broke when the association was preloaded as part of a
larger query that also had additional substitution parameters (for
example, the query in the TotalsReporter). Instead of a sub-query, we
can get the same result (i.e. either the most recent "completed"
submission, or failing that, the most recent submission) with a CASE in
the ORDER clause.