Data migration to fix the state of some submissions
It’s not possible, through application code, to have more than one completed submission on a task.

However, nothing prevents us from messing with the order of operations in the console. When doing manual “replacement returns” for support requests asking to reset tasks, we’ve ended up allowing the old submissions to stay `completed`.

Manual, console interventions will not be necessary in the future.

We now also have the aasm_state ‘replaced’ for old completed submissions.

Co-authored-by: james <james@abscond.org>