Refactor interface to ManagementChargeCalculator::ColumnBased
Rename `column:` parameter on ManagementChargeCalculator::ColumnBased to
`varies_by:`

When thinking about implementing ManagementCharge in FDL, we noticed that the
interfaces to ManagementChargeCalculator::ColumnBased and
ManagementChargeCalculator::FlatRate were confusingly similar.

In ManagementChargeCalculator::ColumnBased, the `column` parameter refers to
the column name which the charge _varies by_.

In ManagementChargeCalculator::FlatRate, the `column` parameter refers to the
column name which the charge itself is taken from.

In ManagementChargeCalculator::ColumnBased, the column name from which the
charge is taken from cannot be varied (it is always the `total_value` attribute
of the entry). So the identical naming of these two parameters between the two
classes was very confusing.

We have renamed `column:` in ManagementChargeCalculator::ColumnBased to make it
obvious this is the factor by which the charge _varies_, not the column the
value is taken from.