Export::Anything#model_classname works with no results
Previously, for invoice and contract exports,
`Export::Anything#model_classname` would use
`@relation.first.entry_type` to determine the type of export that was
running (in order to infer the filename and display the correct message
when no data was exported).

This broke when there were no results, so this change interrogates the
SQL where clause instead.