A known field can't be an additional field
If we miss the type from an Additional field it should be a syntax
error, but was actually being mistakenly parsed as a known field.

If we specify that an additional_field_identifier is .absent? prior
to a known field's identifier, we can short-circuit the parse and
make things stop.