Update database cleaner/transactional fixture conf
Transactional fixtures were enabled in rails_helper.rb, which
contradicted the config in support/database_cleaner.rb. Update the
config there to avoid any ambiguity and ensure transactional fixtures
are definitely disabled.

This commit also updates the DatabaseCleaner config to match the
suggested configuration in the gem's README:

  https://github.com/DatabaseCleaner/database_cleaner#rspec-with-capybara-example

Note that DatabaseCleaner.allow_remote_database_url is set to true so
that the test suite can run on the CI server.