Implement FDL types
We have two types of types: lookups and primitives. Primitives are
things in Field::PRIMITIVE_TYPES (although we repeat ourselves with
the actual literals like "Decimal", "Date" in the parser), and
lookups are assumed to be anything else.

Flesh out field_spec; things were pretty untested before and we were
relying on language_spec as an integration test. We can't do that
any more.