Return error on create submission if task already completed
Currently, when receiving a request to create a new submission for a task where the active submission is already completed, if it's not a correction request, the create submission method responds with the active submission.

This is creating issues in unexpected places, such as when the correction parameter is lost, and the user ends up creating a new file against the completed submission.

We want the API to respond with an error status.

Co-authored-by: james <james@abscond.org>