Make the field_spec more like an integration test
In order to make the additional/known/unknown field types less mysterious, we
have refactored this test to be an integration test.

The `source` strings are actual examples of FDL, instead of AST fragments. This
means this test won't fall out of sync with changes made to the parser and
FDL structure.