Data Warehouse export gets uploaded to S3
Once the export files are generated, they get uploaded to S3 in the
designated bucket (configured using the AWS_S3_EXPORT_BUCKET environment
variable).

The specs for Export::S3Upload make use of the AWS SDKs stub_responses
feature, which allows the API calls to be stubbed and introspected.

We have to disable the Rubocop Style::ClassCheck to get the new rspec
matcher to pass the checks. This is because the Cop tries to enforce
consistent use of `Object#is_a?` or `Object#kind_of?`, which makes no
sense as they can both be useful depending on the circumstances. In this
case, we'd want both a `File` and `Tempfile` to pass this matcher.