Remove export rake tasks / shift spec coverage
Although these rake tasks for exporting individual models are no longer
required, the spec coverage provided important integration-level
coverage of the interaction between Export::Relation and the various
Extract and Row classes for each of the models. The next most sensible
place for these specs to live is the specs on Export::Relation.

This will also make it easier to shift Export::Relation from writing
directly to the /tmp folder generating the exports as Tempfiles.