Correct US date formats in submission entries
Originally, we thought that it all US formatted dates were stored as
'mm/dd/yy' and all UK formatted dates were 'dd/mm/yyyy', so it was
simply a matter of looking for either a 2 or 4 digit year.

Unfortunately, I found several entries that had the date format
'dd/mm/yy', which required cleverer detection. Now, we looking in all
fields which store dates across the *entire* submission to determine
which date parts are contain values greater than 12 (i.e. can't be a
month).

If we can't determine the date format, we rollback all changes the
submission and log the failure.