Switch to using Tempfiles for DW export process
Instead of writing directly to disk, the data warehouse export process
now uses Tempfiles, that way we can leave cleaning up the files after
the export has completed to the Ruby garbage collector. That way we
don't risk our docker containers running out of disk space because of
old exports lingering in /tmp.

This also moves responsibility for naming the export files into the
DataWarehouseExport class, where we can give all the export files for a
given export a consistent timestamp that maps to the export's range_to
value.