Bulk Framework/Supplier import utility code
This codifies what has previously been hand-rolled/copy-pasted into data
migrations when onboarding suppliers for a new framework. Making this a
reusable piece of code a) makes it less likely errors will creep in
when onboarding suppliers, and b) lays the foundation for making this
something that can be performed by an admin user via the web interface.