Remove redundant guard on event
In commit https://github.com/dxw/DataSubmissionServiceAPI/commit/3b9c2c8e82d3a1e4890e702c06f5309bce952401#diff-8b04dfc748ae7c73f82a96d00a3db3c2
a new state, `validation_failed`, was introduced, so the
submission cannot transition to `in_review` if any of its entries have failed validation.

This has moved the need to guard a state transition to earlier in the chain, but we haven’t removed the second guard, on the `reviewed_and_accepted` event. This creates confusion and bugs, so we are now removing the redundant guard.

There is now nothing in the normal operation that would populate the `submission.errors` for the completion action, so that code is redundant as well.