Add and setup Simple Form gem
https://github.com/plataformatec/simple_form

Simple Form allows us to standardise the display of forms across the
application and to use GOVUK patterns.

Along with the gem itself, we've also created some generic error
messages in the locales file and initialised Simple Form with GOVUK
wrapper elements and css class names.

We have also changed the way Simple Form displays required and optional
fields. We are showing a 'This value is optional' hint for optional
fields and nothing (no asterisks) for required fields.