Allow the Rails flashes to have different types
We were only using one flash 'alert' which meant when an operation
was successful the user saw a red status message, which felt like
something had gone wrong.

We've added flashes for each of the visual alerts styles we have:

- failure
- success
- warning
- notice

We can use them for the appropriate actions now with flash[:type] and
the right css class will be applied.