FDL support to compose lookups from other lookups
A few of our Ruby framework defs have "lookups" composed of constants,
e.g.

```
PRIMARY_SPECIALISM_VALUES = (
  CORE_SPECIALISMS +
  NON_CORE_SPECIALISMS
).freeze
```

Add equivalent FDL support, but without the '+' operator.

Example:

```
Lookups {
  One [ '1' ]
  Two [ '2' ]
  Three [ One Two '3' ]
}
```

This will create three lookups:

```
{
  'One'   => ['1'],
  'Two'   => ['2'],
  'Three' => ['1', '2', '3']
}
```

To do this, we needed to start using the `lookup_reference` rule in
two places, once in our new `lookup_item` rule and the existing usage
in `dictionary`'s `allowable_values`. It still needs to behave like a
String but stand out as a LookupReference, so create a String derivative
with no methods so that we can expand references to other lookups
as we encounter them in `AST::Presenter#lookups`.

Co-Authored-By: james <james@abscond.org>