Prevent duplicated entries within a bulk request
We'd previously added a guard to prevent new entries being created when
identical ones already exist in the database. When we changed this to do
bulk insert, we'd forgotten to deduplicate the list of entries we bulk
insert.

This lead to at least one submission containing duplicate rows, which
meant its total value and management charge calculations were out by a
significant amount.

The change removes the duplicates in the array before the bulk insert
loads it into the database.