FDL generator: don't upcase short names
This would attempt to look up a non-existent constant, and we now
have frameworks like RM1043iv with lower-case characters in the
short_name.

While we're about it, update the error message that occurs when
definitions are missing, which were signposting a generator for MISO
CSV we no longer use.