DependentFieldInclusionValidation is case insensitive
On other frameworks, the definition used lower-case values for the
mapping keys. However, on RM3788 the mixed case mapping keys were
causing the DependentFieldInclusionValidation to return `[]` when
performing a lookup, which mean valid combinations of values and parent
values would incorrectly fail validation.

This change downcases all the keys in the hash specified in the `in:`
validator option, which means the lookup works correctly.