Add FDL for framework RM1070
RM1070 is the only framework which had the 'Quantity' field ingested as an
integer. All other frameworks have this as a decimal.

We have changed this framework to ingest the field as a decimal to bring it
in line with other frameworks. If this causes issues down the line, we can
revisit.