Fix FDL presence handling inconsistencies
We had previously assumed that the two `inclusion`-type validators were
stricter than (and obviated) presence, but in fact the two need to work
together.

Fix this by leaving presence: true on YesNo and lookup fields, and
alter the FDL of failing framework definitions to match this (some of
our fields that we had assumed were covered for presence by these
inclusion validators were not).