Data migration for updating  the invalid data fields
NB: Due to the large size of the CSV containing the correction data, it
has *NOT* been commited here. It can instead be downloaded ahead of the
migration being run.

See the `wip/973-american-date-formats` branch for the code which
generated the CSV (not included here as it required 'roo' and 'roo-xls')