Style an FDL parse failure
Use the Figma designs to style. See

https://www.figma.com/proto/JbUKCvCtcFteaOT8AfCkCxHn/fdl-tooling?node-id=3%3A3632&scaling=min-zoom&redirected=1

This means we need a special :fdl_failure flash type, as it's the
only way to communicate to the _alert partial that we want different
classes on things. Extend the normal failure class and pre-wrap the
Parslet parse tree failures.

Add a "Definition Error" header when the flash is :fdl_failure.