Use standard AR model errors for FDL
We were using flash[:fdl_failure], because we were able to easily
generate a definition error of a style that corresponded to the Figma
designs.

Unfortunately this made things inconsistent with all our other errors.
Accept the cost of triggering a parse twice (once for errors via an
FdlValidator and once for actually getting the details) and remove the
calls to flash[].

Also, style the errors for FDL definition_source problems only to remove
the bold and use pre-formatted text so we can look at parse tree
problems sensibly.