Ensure that we can read XLS files generated by LibreOffice
There was a bug in the old Lambda-based ingest that mean it couldn't
correctly retrieve the sheet names from XLS files generated by
LibreOffice, due to difference in the way this metadata was stored.

I've made the `sheets` method public in order to test that the sheet
names are retrievable by new ingest, without having to fully convert the
files, which slows down the test suite.