Favour require_dependency for 'framework'
According to autoload docs, we should never explicitly `require` any
class that is subject to autoload, which in practice means anything
outside of `./lib`.

https://guides.rubyonrails.org/autoloading_and_reloading_constants.html#autoloading-and-require

We do, however, need to ensure that app/models/framework has been
loaded first to avoid superclass mismatch errors later. The preferred
mechanism for this is `require_dependency`.