Add a Framework::Definition::Generator
Replaces a lot of Language. It's main purpose is to track the last error
in class generation. That can either be a parse error – a full
preformatted ASCII tree – or the last semantic error the Transpiler
encountered. At present, the only semantic error is the non-occurrence
of an InvoiceValue field in InvoiceFields or ContractValue in
ContractFields.

The generator takes some source and optionally a logger. You can ask it
if it was a #success? or an #error? and in the former case a #definition
will be available. In the case of an #error? it will be nil.