Convert integer date fields (Excel epoch) to dd/mm/yyyy format
When spreadsheets contain a mixture of date formats, it causes csvkit to
get confused, so some cells are output as their raw values instead of in
ISO8601. This is the number of days since 1/1/1900, the Excel epoch.

NB: There's a slight risk that dates will be off by 4 years, but this
requires a specific buggy version of Excel for Macintosh which is quite
old at this point.