Treat booleans contained in numeric fields correctly
Sometimes csvkit's type detector incorrectly reads a column containing
all zeroes or ones as a boolean, which results in '0' and '1' being
converted to 'False' and 'True' respectively.

We were previously mapping these values to 'N' and 'Y' as per CCS's
requirement for booleans.

This change ensures that numeric fields (those that get an
IngestedNumericalityValidator applied) get those booleans correctly
mapped to `0` and `1` instead.