Make admin submission file downloads compatible with GPaaS
Previously, it was using a method that would redirect the user to a
generated temporary URL for the file in the S3 bucket. Since GPaaS
does not permit this, we need to fetch the file ourselves and stream it
back to the user.

As a bonus, the files are returned with filenames that are more useful
to admin users, who likely have many downloads from varying suppliers
and frameworks - e.g. "RM1234 Company Name Ltd (December 2018).xls".