Ingest fetches files using ActiveStorage, for GPaaS compatibility
GPaaS doesn't allow files to be downloaded from a S3 bucket using a
signed public URL, but instead needs to be fetched using the AWS S3
APIs.

NB: In Rails 5.2, an S3-specific exception is raised in the event of a
failure fetching the file. This makes it hard to test, as this is
provider-specific, and the test ActiveStorage provider does not have
this. Rails 6 will have ActiveStorage::FileNotFoundError.

For now, I've stubbed the ActiveStorage `download` method to raise the
expected AWS exception.