Expose the key of the framework's template file key on task
To enable suppliers to be able to download the template file for a task, we need to expose the `key` of the activestorage file object.

This is because the Supplier Facing App does not have access to the ActiveStorage object over the API. By giving the `key`, which is the filename stored in S3, the supplier facing app is able to use the S3 library directly.

The supplier facing app will be unable to point to a public URL for the S3 file, as that is not supported on GPAAS S3, which we will be moving to soon.

To ensure that we don't make needless requests against the ActiveStorage tables on *all* API calls that include a task or framework, I have added it as an attribute on SerializableTask, which enables use to only fetch the key if @include_file is exposed, which we only do in the relevant API request. Ideally it would be an attribute on the included Framework, but JSONAPI does not allow us to pass exposed variables to included associations. Thanks to @leeky for figuring this out.