Add a manifest for a sidekiq workers
We use sidekiq to run background tasks such as ingesting submissions.
This currently needs to be a separate app so create a new manifest for
it.
Of note we need to not map a route and change the health check to
process when compared to a normal application on the PaaS