Add #load_lots! to Framework
This method reads the FDL source, and ensures that the database reflects the lots.

At the moment, we don't really know how we will support attempts by admins to delete a lot that has agreements against it, and for the time being should be impossible as an admin cannot edit a published Framework.