Import customers from attached Excel file
The import process happens in an ActiveJob worker that is fired as soon
as a URN list is finished uploading.

Importing fails if the Excel file doesn't contain a sheet named
'Customers', or that sheet doesn't contain the required columns - URN,
CustomerName, PostCode and Sector.

If all is ok, we perform a PostgreSQL upsert, which inserts new
customers, and updates any that have the same URN.