Add management charge to supplier's task view
In order to make it possible to test submissions against new frameworks,
we need to show the total spend and management charge to admins, so they
can double check the calculations.

This meant we had to reorganise the supplier's task list in order to fit
the additional columns.

Firstly, I merged "Framework" and "Month" columns and renamed it to
"Task". Then I renamed the "Download submission file" link to just
"Download".